core code updates
git-svn-id: https://svn.code.sf.net/p/hsqldb/svn/base/trunk@2922 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/result/ResultMetaData.java
@@ -220,7 +220,7 @@ public final class ResultMetaData {
     }
 
     private static void decodeTableColumnAttrs(int in, ColumnBase column) {
-        column.setNullability(in & 0x0000000f);
+        column.setNullability((byte) (in & 0x0000000f));
         column.setIdentity((in & 0x00000010) != 0);
     }
